Ayushk1804 Solution
Approach:

Split the string to get an array of four strings i.e. A, B, C, D.
If any one of them contains ?, then for each position check whether the required number and given number match or not (excluding the ? position).
If they match, then return the missing digit in the given number (position of ? in given number).
I've completed the given assignment and tested with the given testcases and some extra edge cases. I've added my solution with this PR.